// frontend/packages-test/workflow/base/src/types/dto.ts
export namespace DTODefine {
  export type LiteralExpressionContent =
    | string
    | number
    | boolean
    | Array<unknown>;

  export interface LiteralExpression {
    type: 'literal';
    content?: LiteralExpressionContent;
    rawMeta?: ValueExpressionRawMeta;
  }

  export type RefExpressionContent =
    | {
        source: 'variable';
        blockID: undefined;
        name: string;
      }
    | {
        source: 'block-output';
        blockID: string;
        name: string;
      }
    | {
        source: `global_variable_${string}`;
        path: string[];
        blockID: string;
        name: string;
      };
  export interface RefExpression {
    type: 'ref';
    content?: RefExpressionContent;
    rawMeta?: ValueExpressionRawMeta;
  }

  export interface ObjectRefExpression {
    type: 'object_ref';
    content?: unknown;
    rawMeta?: ValueExpressionRawMeta;
  }

  // 当 schema 为 string 时表示原始的 FDL 类型export type ListVariableSchema = VariableTypeDef & Omit<VariableOption, 'name'>
  export type ObjectVariableSchema = InputVariableDTO[];
  export type ListVariableSchema = VariableTypeDef &
    Omit<VariableOption, 'name'>;
  export type VariableSchema = ListVariableSchema | ObjectVariableSchema;
  export type BasicVariableType =
    | 'string'
    | 'integer'
    | 'float'
    | 'boolean'
    | 'image'
    | 'unknown';
  export type ComplexVariableType = 'object' | 'list';
  export type VariableType = BasicVariableType | ComplexVariableType;

  export interface BasicVarTypeDef {
    type: BasicVariableType;
    assistType?: AssistTypeDTO;
  }
  export interface ObjectVarTypeDef {
    type: VariableTypeDTO.object;
    schema?: ObjectVariableSchema;
    assistType?: AssistTypeDTO;
  }
  export interface ListVarTypeDef {
    type: VariableTypeDTO.list;
    schema: ListVariableSchema;
    assistType?: AssistTypeDTO;
  }

  export type VariableTypeDef =
    | BasicVarTypeDef
    | ObjectVarTypeDef
    | ListVarTypeDef;
  interface VariableOption {
    name: string;
    label?: string;
    defaultValue?: LiteralExpressionContent;
    description?: string;
    required?: boolean;
  }

  export type InputVariableDTO = VariableOption & VariableTypeDef;
}
// ...
export interface ValueExpressionDTO {
  type?: string;
  assistType?: number;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  schema?: any;
  value:
    | DTODefine.LiteralExpression
    | DTODefine.RefExpression
    | DTODefine.ObjectRefExpression;
}

// eslint-disable-next-line @typescript-eslint/no-namespace
export namespace ValueExpressionDTO {
  /**
   * 空引用
   */
  export function createEmpty(): ValueExpressionDTO {
    return {
      type: 'string',
      value: {
        type: 'ref',
        content: {
          source: 'block-output',
          blockID: '',
          name: '',
        },
      },
    };
  }
}
export interface InputValueDTO {
  name?: string;
  input: ValueExpressionDTO;
  schema?: InputValueDTO[];
  id?: string;
}

export enum VariableTypeDTO {
  object = 'object',
  list = 'list',
  string = 'string',
  integer = 'integer',
  float = 'float',
  boolean = 'boolean',
  image = 'image',
  time = 'time',
}

export enum AssistTypeDTO {
  file = 1,
  image = 2,
  doc = 3,
  code = 4,
  ppt = 5,
  txt = 6,
  excel = 7,
  audio = 8,
  zip = 9,
  video = 10,
  svg = 11,
  voice = 12,
  time = 10000,
}